requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
alembic==1.4.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
databases==0.2.6
idna==2.9
kafka-python==2.0.1
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
multidict==4.7.5
orm==0.1.5
pika==1.1.0
pkg-resources==0.0.0
pprint==0.1
pycodestyle==2.5.0
pycparser==2.20
pymongo==3.10.1
PyMySQL==0.9.2
pynsq==0.8.3
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
tornado==4.5.3
typesystem==0.2.4
tzlocal==2.0.0
urllib3==1.25.8
yarl==1.4.2
